Teach mips_syscall how to pass parameters to guest program
Previously, the MIPS model relied on a simple program startup model
that assumed the guest program was linked with a particular ArchC
library for guests.

Since we are changing this model to accept any MIPSr32 binary, program
start must follow ABI rules. This patch also introduces an auxv entry
to signal the guest program that the system page size is 4096, which is
necessary for the malloc implementation to work properly on programs
compiled with ecc (ellcc compiler).

This patch also fixes encodings for shift instructions, previously only
partially defined.