Override the _exit syscall to return correct val
The default implementation of the _exit syscall (ac_syscall.H) returns
the value from the `get_int(0)`. The `get_int` in MIPS model returns
RB[argn+4]. So, the register R4 should contain the return value.
But the MIPS cross-compiler is putting the return value in R2. So,
I overrided the _exit syscall to return the R2 value.